partually integrated
integrated the grouping in the control method, but it is only passing tht last group to d3ngStackedenduse. I need to update it to pass in an array of objects, and hten have that method loop through the array to make mulitple charts.

If not easy to add multiple charts this way, maybe there is a way to update the html section to loop through each scenario set.